requirements.txt
fastapi
uvicorn
openai>=1.0.0
python-multipart
pdfplumber
python-docx
python-dotenv
